feature: resolve compiler and runtime errors due to merge.
