+++ app/src/main/java/com/amsterdam/aflami/di/localDataSourceModule.kt
@@ -6,12 +6,16 @@ import com.example.localdatasource.roomDataBase.daos.CountryDao
 import com.example.localdatasource.roomDataBase.daos.MovieDao
 import com.example.localdatasource.roomDataBase.daos.RecentSearchDao
 import com.example.localdatasource.roomDataBase.daos.TvShowDao
+import com.example.localdatasource.roomDataBase.datasource.LocalCategoryDataSourceImpl
 import com.example.localdatasource.roomDataBase.datasource.LocalCountryDataSourceImpl
 import com.example.localdatasource.roomDataBase.datasource.LocalMovieDataSourceImpl
 import com.example.localdatasource.roomDataBase.datasource.RecentSearchDataSourceImpl
+import com.example.remotedatasource.datasource.RemoteCategoryDatasourceImpl
+import com.example.repository.datasource.local.LocalCategoryDataSource
 import com.example.repository.datasource.local.LocalCountryDataSource
 import com.example.repository.datasource.local.LocalMovieDataSource
 import com.example.repository.datasource.local.LocalRecentSearchDataSource
+import com.example.repository.datasource.remote.RemoteCategoryDatasource
 import org.koin.android.ext.koin.androidApplication
 import org.koin.dsl.module
 
@@ -27,4 +31,6 @@ val localDataSourceModule = module {
     single<LocalCountryDataSource> { LocalCountryDataSourceImpl(get()) }
     single<LocalMovieDataSource> { LocalMovieDataSourceImpl(get()) }
     single<LocalRecentSearchDataSource> { RecentSearchDataSourceImpl(get()) }
+    single<RemoteCategoryDatasource> { RemoteCategoryDatasourceImpl(get()) }
+    single<LocalCategoryDataSource> { LocalCategoryDataSourceImpl(get()) }
 }
