+++ app/src/main/java/com/amsterdam/aflami/di/repositoryModule.kt
@@ -1,14 +1,21 @@
 package com.amsterdam.aflami.di
 
+import com.example.domain.repository.CategoryRepository
 import com.example.domain.repository.CountryRepository
 import com.example.domain.repository.MovieRepository
+import com.example.domain.repository.RecentSearchRepository
+import com.example.domain.repository.TvShowRepository
 import com.example.repository.mapper.local.CategoryLocalMapper
 import com.example.repository.mapper.local.CountryLocalMapper
 import com.example.repository.mapper.local.MovieLocalMapper
+import com.example.repository.mapper.local.RecentSearchMapper
 import com.example.repository.mapper.remote.RemoteCountryMapper
 import com.example.repository.mapper.remote.RemoteMovieMapper
+import com.example.repository.repository.CategoryRepositoryImpl
 import com.example.repository.repository.CountryRepositoryImpl
 import com.example.repository.repository.MovieRepositoryImpl
+import com.example.repository.repository.RecentSearchRepositoryImpl
+import com.example.repository.repository.TvShowRepositoryImpl
 import org.koin.dsl.module
 
 val repositoryModule = module {
@@ -17,6 +24,10 @@ val repositoryModule = module {
     single { CategoryLocalMapper() }
     single { MovieLocalMapper(get()) }
     single { RemoteMovieMapper() }
+    single { RecentSearchMapper() }
     single<CountryRepository> { CountryRepositoryImpl(get(), get(), get(), get()) }
-    single<MovieRepository> { MovieRepositoryImpl(get(), get(), get(), get(), get(), get()) }
+    single<MovieRepository> { MovieRepositoryImpl(get(), get(), get(), get(), get()) }
+    single<CategoryRepository> { CategoryRepositoryImpl(get(), get(), get()) }
+    single<RecentSearchRepository> { RecentSearchRepositoryImpl(get(), get()) }
+    single<TvShowRepository> { TvShowRepositoryImpl() }
 }
