+++ app/src/main/java/com/amsterdam/aflami/di/useCaceModule.kt
@@ -1,5 +1,7 @@
 package com.amsterdam.aflami.di
 
+import com.example.domain.useCase.GetMovieCategoriesUseCase
+import com.example.domain.useCase.GetMoviesByActorUseCase
 import com.example.domain.useCase.GetMoviesByKeywordUseCase
 import com.example.domain.useCase.GetTvShowByKeywordUseCase
 import com.example.domain.useCase.search.ClearAllRecentSearchesUseCase
@@ -8,6 +10,7 @@ import com.example.domain.useCase.search.GetRecentSearchesUseCase
 import org.koin.core.module.dsl.singleOf
 import com.example.domain.useCase.GetMoviesByCountryUseCase
 import com.example.domain.useCase.GetSuggestedCountriesUseCase
+import com.example.domain.useCase.GetTvShowCategoriesUseCase
 import com.example.domain.validation.CountryValidator
 import com.example.domain.validation.CountryValidatorImp
 import org.koin.dsl.module
@@ -21,4 +24,7 @@ val useCaseModule = module{
     single<CountryValidator> { CountryValidatorImp() }
     single { GetSuggestedCountriesUseCase(get(), get()) }
     single { GetMoviesByCountryUseCase(get()) }
+    single { GetMoviesByActorUseCase(get()) }
+    single { GetMovieCategoriesUseCase(get()) }
+    single { GetTvShowCategoriesUseCase(get()) }
 }
