+++ designSystem/src/main/java/com/example/designsystem/DesignSystem.kt
@@ -223,7 +223,7 @@ fun DesignSystem(){
             )
         }
         MovieCard(
-            movieImage = painterResource(R.drawable.bg_man_with_popcorn),
+            movieImage = "",
             movieTitle = "Eternal Sunshine of the Spotless mind",
             movieType = "Sci-fi",
             movieYear = "2002",
@@ -231,7 +231,7 @@ fun DesignSystem(){
             movieRating = "9.8"
         )
         UpcomingCard(
-            movieImage = painterResource(R.drawable.bg_man_with_popcorn),
+            movieImage = "",
             movieTitle = "Eternal Sunshine of the Spotless mind",
             movieType = "Sci-fi",
             movieYear = "2002",
