+++ designSystem/src/main/java/com/example/designsystem/components/UpcomingCard.kt
@@ -13,7 +13,7 @@ import com.example.designsystem.utils.ThemeAndLocalePreviews
 
 @Composable
 fun UpcomingCard(
-    movieImage: Painter,
+    movieImage: String,
     movieTitle: String,
     movieType: String,
     movieYear: String,
@@ -41,7 +41,7 @@ fun UpcomingCard(
 private fun UpcomingCardPreview() {
     AflamiTheme {
         UpcomingCard(
-            movieImage = painterResource(R.drawable.bg_children_wearing_3d),
+            movieImage = "",
             movieType = "TV show",
             movieYear = "2016",
             movieTitle = "Your Name",
