+++ localDatasource/src/main/java/com/example/localdatasource/roomDataBase/daos/TvShowDao.kt
@@ -19,7 +19,7 @@ interface TvShowDao {
     @Query(
         """
         SELECT * FROM tv_shows 
-        WHERE id IN (
+        WHERE categoryId IN (
             SELECT tvShowId FROM tvshow_search
             WHERE searchKeyword = :keyword
         )
