+++ ui/src/main/java/com/example/ui/screens/search/SearchScreen.kt
@@ -1,70 +1,145 @@
 package com.example.ui.screens.search
 
+import android.util.Log
+import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.statusBarsPadding
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import com.example.designsystem.components.buttons.PrimaryButton
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.example.designsystem.R
+import com.example.designsystem.components.MovieCard
+import com.example.designsystem.components.TabsLayout
+import com.example.designsystem.components.TextField
+import com.example.designsystem.components.appBar.DefaultAppBar
 import com.example.designsystem.theme.AppTheme
-import com.example.ui.application.LocalNavController
-import com.example.ui.navigation.Route
+import com.example.ui.screens.search.sections.RecentSearchesSection
+import com.example.ui.screens.search.sections.SuggestionsHubSection
+import com.example.viewmodel.common.MediaType
+import com.example.viewmodel.common.TabOption
+import com.example.viewmodel.search.GlobalSearchInteractionListener
+import com.example.viewmodel.search.GlobalSearchViewModel
+import com.example.viewmodel.search.SearchUiEffect
+import com.example.viewmodel.search.SearchUiState
+import kotlinx.coroutines.flow.collectLatest
+import org.koin.androidx.compose.koinViewModel
 
 @Composable
 fun SearchScreen(
-    modifier: Modifier = Modifier
+    viewModel: GlobalSearchViewModel = koinViewModel(),
 ) {
-    val navController = LocalNavController.current
-    SearchScreenContent(
-        modifier = modifier,
-        onSearchByActorClicked = { navController.navigate(Route.SearchByActor) },
-        onSearchByCountryClicked = { navController.navigate(Route.SearchByCountry) },
-    )
-}
+    val state by viewModel.state.collectAsState()
 
-@Composable
-private fun SearchScreenContent(
-    modifier: Modifier = Modifier,
-    onSearchByActorClicked: () -> Unit,
-    onSearchByCountryClicked: () -> Unit
-) {
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .background(AppTheme.color.surface),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Row {
-            PrimaryButton(
-                modifier = Modifier.weight(1f),
-                title = "Country",
-                onClick = onSearchByCountryClicked,
-                isEnabled = true,
-                isLoading = false,
-                isNegative = false
-            )
-            PrimaryButton(
-                modifier = Modifier.weight(1f),
-                title = "Actor",
-                onClick = onSearchByActorClicked,
-                isEnabled = true,
-                isLoading = false,
-                isNegative = false
-            )
+    LaunchedEffect(Unit) {
+        viewModel.effect.collectLatest { effect ->
+            when (effect) {
+                SearchUiEffect.NavigateBack -> {
+                    Log.e("bk", "collect: NavigateBack")
+                }
+
+                SearchUiEffect.NavigateToActorSearch -> {
+                    Log.e("bk", "collect: NavigateToActorSearch")
+                }
+
+                SearchUiEffect.NavigateToMovieDetails -> {
+                    Log.e("bk", "collect: NavigateToMovieDetails")
+                }
+
+                SearchUiEffect.NavigateToWorldSearch -> {
+                    Log.e("bk", "collect: NavigateToWorldSearch")
+                }
+
+                else -> {}
+            }
         }
     }
+
+    SearchContent(state = state, interaction = viewModel)
 }
 
-@Preview
 @Composable
-private fun SearchScreenPreview() {
-    SearchScreenContent(
-        onSearchByActorClicked = {},
-        onSearchByCountryClicked = {}
-    )
+private fun SearchContent(state: SearchUiState, interaction: GlobalSearchInteractionListener) {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(color = AppTheme.color.surface)
+            .statusBarsPadding()
+    ) {
+        DefaultAppBar(
+            modifier = Modifier.padding(horizontal = 16.dp),
+            title = stringResource(R.string.search),
+            onNavigateBackClicked = interaction::onNavigateBackClicked
+        )
+
+        TextField(
+            modifier = Modifier
+                .background(color = AppTheme.color.surface)
+                .padding(top = 8.dp)
+                .padding(horizontal = 16.dp),
+            text = state.query,
+            onValueChange = interaction::onTextValuedChanged,
+            hintText = stringResource(R.string.search_hint),
+            trailingIcon = R.drawable.ic_filter_vertical,
+            onTrailingClick = interaction::onFilterButtonClicked,
+            isError = state.errorUiState != null,
+            errorMessage = getErrorMessageBySearchErrorUiState(state.errorUiState),
+            maxCharacters = 100
+        )
+
+        AnimatedVisibility(state.query.isNotEmpty()) {
+            LazyVerticalGrid(
+                modifier = Modifier.height((state.movies.count() * 222).dp),
+                columns = GridCells.Adaptive(160.dp),
+                verticalArrangement = Arrangement.spacedBy(8.dp),
+                horizontalArrangement = Arrangement.spacedBy(8.dp),
+                contentPadding = PaddingValues(horizontal = 16.dp, vertical = 12.dp)
+            ) {
+                stickyHeader {
+                    TabsLayout(
+                        modifier = Modifier.fillMaxWidth(),
+                        tabs = listOf(
+                            stringResource(R.string.movies),
+                            stringResource(R.string.tv_shows)
+                        ),
+                        selectedIndex = state.selectedTabOption.index,
+                        onSelectTab = { index -> interaction.onTabOptionClicked(TabOption.entries[index]) },
+                    )
+                }
+                items(
+                    if (state.selectedTabOption == TabOption.MOVIES) state.movies else state.tvShows,
+                ) { mediaItem ->
+                    mediaItem.apply {
+                        MovieCard(
+                            movieImage = "",
+                            movieType = if (mediaType == MediaType.TV_SHOW) stringResource(R.string.tv_shows)
+                            else stringResource(R.string.movies),
+                            movieYear = yearOfRelease,
+                            movieTitle = name,
+                            movieRating = rate,
+                        )
+                    }
+                }
+            }
+
+        }
+
+        SuggestionsHubSection(state = state, interaction = interaction)
+
+        RecentSearchesSection(state = state, interaction = interaction)
+
+    }
 }
