+++ ui/src/main/java/com/example/ui/screens/search/sections/filterDialog/FilterDialog.kt
@@ -1,4 +1,4 @@
-package com.example.ui.screen.search.sections.filterDialog
+package com.example.ui.screens.search.sections.filterDialog
 
 
 import androidx.compose.foundation.background
