+++ viewModel/src/main/java/com/example/viewmodel/searchByActor/SearchByActorScreenState.kt
@@ -1,0 +1,9 @@
+package com.example.viewmodel.searchByActor
+
+import com.example.viewmodel.search.countrySearch.MovieUiState
+
+data class SearchByActorScreenState(
+    val isLoading:Boolean = false,
+    val query:String="",
+    val movies:List<MovieUiState> = emptyList(),
+)
